--- delivery-service/src/main/java/com/example/deliveryservice/dto/LandmarkDto.java
@@ -1,16 +1,0 @@
-package com.example.deliveryservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LandmarkDto {
-    private Integer id;
-    private String name;
-    private Integer cityId;
-}
