--- delivery-service/src/main/java/com/example/deliveryservice/dto/RegionDto.java
@@ -1,18 +1,0 @@
-package com.example.deliveryservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class RegionDto {
-    private Integer id;
-    private String name;
-    private List<CityDto> cities;
-}
