--- delivery-service/src/main/java/com/example/deliveryservice/entity/PromoCode.java
@@ -1,31 +1,0 @@
-package com.example.deliveryservice.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.util.Date;
-
-@Entity
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "promo_codes")
-public class PromoCode {
-
-    @Id
-    @GeneratedValue(generator = "promo_code_id")
-    @SequenceGenerator(name = "promo_code_id", sequenceName = "promo_code_id_seq", allocationSize = 1)
-    private Integer id;
-
-    private String name;
-
-    @Column(name = "start_date")
-    private Date startDate;
-
-    @Column(name = "end_date")
-    private Date endDate;
-}
