--- delivery-service/src/main/java/com/example/deliveryservice/repository/CityRepository.java
@@ -1,13 +1,0 @@
-package com.example.deliveryservice.repository;
-
-import com.example.deliveryservice.entity.City;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface CityRepository extends JpaRepository<City, Integer> {
-
-    List<City> findAllByRegionId(Integer id);
-}
