--- delivery-service/src/main/java/com/example/deliveryservice/repository/RegionRepository.java
@@ -1,9 +1,0 @@
-package com.example.deliveryservice.repository;
-
-import com.example.deliveryservice.entity.Region;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RegionRepository extends JpaRepository<Region, Integer> {
-}
