--- delivery-service/src/main/java/com/example/deliveryservice/service/CityService.java
@@ -1,13 +1,0 @@
-package com.example.deliveryservice.service;
-
-import com.example.deliveryservice.dto.CityDto;
-import org.springframework.stereotype.Service;
-import shared.libs.dto.ResponseDto;
-
-import java.util.List;
-
-@Service
-public interface CityService {
-
-    ResponseDto<List<CityDto>> getAllByRegionId(Integer id);
-}
