--- delivery-service/src/main/java/com/example/deliveryservice/service/mapper/CityMap.java
@@ -1,37 +1,0 @@
-package com.example.deliveryservice.service.mapper;
-
-import com.example.deliveryservice.dto.CityDto;
-import com.example.deliveryservice.entity.City;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class CityMap {
-
-    public static List<CityDto> toDtoList(List<City> cities) {
-        return cities.stream().map(CityMap::toDtoWithOutLandmark).collect(Collectors.toList());
-    }
-
-    public static CityDto toDtoWithOutLandmark(City city) {
-        return CityDto.builder()
-                .id(city.getId())
-                .name(city.getName())
-                .build();
-    }
-
-    public static CityDto toDto(City city) {
-        return CityDto.builder()
-                .id(city.getId())
-                .name(city.getName())
-                .landmarks(LandmarkMap.toDtoList(city.getLandmarks()))
-                .build();
-    }
-
-    public static City toEntity(CityDto cityDto) {
-        return City.builder()
-                .id(cityDto.getId())
-                .name(cityDto.getName())
-                .regionId(cityDto.getRegionId())
-                .build();
-    }
-}
