--- delivery-service/src/main/java/com/example/deliveryservice/service/mapper/RegionMap.java
@@ -1,29 +1,0 @@
-package com.example.deliveryservice.service.mapper;
-
-import com.example.deliveryservice.dto.RegionDto;
-import com.example.deliveryservice.entity.Region;
-
-public class RegionMap {
-
-    public static RegionDto toDtoWithOutCity(Region region) {
-        return RegionDto.builder()
-                .id(region.getId())
-                .name(region.getName())
-                .build();
-    }
-
-    public static RegionDto toDto(Region region) {
-        return RegionDto.builder()
-                .id(region.getId())
-                .name(region.getName())
-                .cities(CityMap.toDtoList(region.getCities()))
-                .build();
-    }
-
-    public static Region toEntity(RegionDto regionDto) {
-        return Region.builder()
-                .id(regionDto.getId())
-                .name(regionDto.getName())
-                .build();
-    }
-}
