+++ delivery-service/src/main/java/uz/nt/deliveryservice/dto/CityDto.java
@@ -1,0 +1,19 @@
+package uz.nt.deliveryservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CityDto {
+    private Integer id;
+    private String name;
+    private Integer regionId;
+    private List<LandmarkDto> landmarks;
+}
