+++ delivery-service/src/main/java/uz/nt/deliveryservice/dto/LandmarkDto.java
@@ -1,0 +1,16 @@
+package uz.nt.deliveryservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class LandmarkDto {
+    private Integer id;
+    private String name;
+    private Integer cityId;
+}
