+++ delivery-service/src/main/java/uz/nt/deliveryservice/dto/PromoCodeDto.java
@@ -1,0 +1,19 @@
+package uz.nt.deliveryservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Date;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PromoCodeDto {
+    private Integer id;
+    private String name;
+    private Date startDate;
+    private Date endDate;
+}
