+++ delivery-service/src/main/java/uz/nt/deliveryservice/repository/PromoCodeRepository.java
@@ -1,0 +1,9 @@
+package uz.nt.deliveryservice.repository;
+
+import uz.nt.deliveryservice.entity.PromoCode;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PromoCodeRepository extends JpaRepository<PromoCode, Integer> {
+}
